LayoutTests/platform/mac/accessibility/css-content-attribute-expected.txt

SAMPLE

This tests that when the content attribute in CSS is used, the correct values are returned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXStaticText'
PASS body.childAtIndex(0).childAtIndex(1).role is 'AXRole: AXStaticText'
PASS body.childAtIndex(0).childAtIndex(2).role is 'AXRole: AXStaticText'
PASS body.childAtIndex(0).childAtIndex(0).stringValue is 'AXValue: This is a'
PASS body.childAtIndex(0).childAtIndex(1).stringValue is 'AXValue: SAMPLE'
PASS body.childAtIndex(0).childAtIndex(2).stringValue is 'AXValue: Headline'
PASS successfullyParsed is true

TEST COMPLETE
